requirements.txt
absl-py==0.10.0
astunparse==1.6.3
backcall==0.2.0
blis==0.4.1
cachetools==4.1.1
catalogue==1.0.0
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
configparser==5.0.0
cymem==2.0.3
decorator==4.4.2
docker-pycreds==0.4.0
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-2.3.1/en_core_web_sm-2.3.1.tar.gz
filelock==3.0.12
future==0.18.2
gast==0.3.3
gitdb==4.0.5
GitPython==3.1.7
google-auth==1.21.0
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
gql==0.2.0
graphql-core==1.1
grpcio==1.31.0
h5py==2.10.0
idna==2.10
jedi==0.17.2
joblib==0.16.0
Keras==2.4.3
Keras-Preprocessing==1.1.2
Markdown==3.2.2
murmurhash==1.0.2
numpy==1.18.5
nvidia-ml-py3==7.352.0
oauthlib==3.1.0
opt-einsum==3.3.0
packaging==20.4
pandas==1.1.1
parso==0.7.1
pathtools==0.1.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.2.0
pip-tools==5.3.1
plac==1.1.3
preshed==3.0.2
promise==2.3
prompt-toolkit==3.0.7
protobuf==3.13.0
psutil==5.7.2
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.6.1
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.1
PyYAML==5.3.1
regex==2020.7.14
requests==2.24.0
requests-oauthlib==1.3.0
rouge==1.0.0
rouge_metric==1.0.1
rsa==4.6
sacremoses==0.0.43
scikit-learn==0.23.2
scipy==1.4.1
sentencepiece==0.1.91
sentry-sdk==0.17.2
shortuuid==1.0.1
six==1.15.0
sklearn==0.0
smmap==3.0.4
spacy==2.3.2
srsly==1.0.2
subprocess32==3.5.4
tensorboard==2.3.0
tensorboard-plugin-wit==1.7.0
tensorflow==2.3.0
tensorflow-estimator==2.3.0
termcolor==1.1.0
thinc==7.4.1
threadpoolctl==2.1.0
tokenizers==0.8.1rc2
torchtext==0.7.0
tqdm==4.48.2
traitlets
transformers==3.1.0
urllib3==1.25.10
wandb==0.9.6
wasabi==0.8.0
watchdog==0.10.3
wcwidth==0.2.5
Werkzeug==1.0.1
wrapt==1.12.1
nltk